Fix makefile
